requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.6
django-phonenumber-field[phonenumberslite]==7.1.0
coreapi==2.3.3
